AX: [GTK] Regression: accessibility/gtk/aria-keyshortcuts.html fails after r224953 https://bugs.webkit.org/show_bug.cgi?id=179824
Reviewed by Chris Fleizach.

The failure was due to AXPropertyName::KeyShortcuts not being included in
the group of properties whose value is a string. As a result, the new method
to get string values, AccessibleNode::effectiveStringValueForElement(), was
returning an empty String. This was not caught by the shared set of tests
because at the present time, only WebKitGtk supports aria-keyshortcuts.

No new tests needed. The regression was caught by an existing test.

* accessibility/AccessibleNode.cpp:
(WebCore::isPropertyValueString):


Identifier: 195842@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224970 268f45cc-cd09-0410-ab3c-d52691b4dbfc