[CoordGraphics] Remove relay CoordinatedBuffer::Client implementations https://bugs.webkit.org/show_bug.cgi?id=179970
Reviewed by Carlos Garcia Campos.

Source/WebCore:

Drop CoordinatedBuffer::painToSurface() in favor of a simpler
GraphicsContext reference getter that returns a reference to the
GraphicsContext object of the internal ImageBuffer.

Remove the ImageBackingSurfaceClient class that only acted as a relay
that enabled a CoordinatedBuffer to draw an image through its context.
Instead, we use the GraphicsContext of the allocated CoordinatedBuffer
to draw the image directly. We still use a rectangle whose size
matches the size of the given Image as the clipping area and as the
source and destination rectangle for the image.

No new tests -- no change in behavior.

* platform/graphics/texmap/coordinated/CoordinatedBuffer.cpp:
(WebCore::CoordinatedBuffer::context):
(WebCore::CoordinatedBuffer::paintToSurface): Deleted.
* platform/graphics/texmap/coordinated/CoordinatedBuffer.h:
* platform/graphics/texmap/coordinated/CoordinatedImageBacking.cpp:
(WebCore::CoordinatedImageBacking::update):
(WebCore::ImageBackingSurfaceClient::ImageBackingSurfaceClient): Deleted.

Source/WebKit:

Drop the helper UpdateAtlasSurfaceClient class and instead directly use
the CoordinatedBuffer's GraphicsContext to perform clip and translation
operations based on the allocated rectangle, to pre-fill the target rect
with transparent color in case the alpha channel is supported, and to
then invoke CoordinatedBuffer::Client::paintToSurfaceContext() method,
passing the GraphicsContext as the argument.

* WebProcess/WebPage/CoordinatedGraphics/UpdateAtlas.cpp:
(WebKit::UpdateAtlas::paintOnAvailableBuffer):
(): Deleted.


Identifier: 195981@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225113 268f45cc-cd09-0410-ab3c-d52691b4dbfc