[CoordGraphics] UpdateAtlas constructor should receive an IntSize, not a dimension value https://bugs.webkit.org/show_bug.cgi?id=179976
Reviewed by Carlos Garcia Campos.

Have the UpdateAtlas constructor receive an IntSize object that specifies
the desired size for the CoordinatedBuffer. Passing in a dimension doesn't
really make sense since this value isn't dynamically configurable, and
the only value that is passed in is already a power-of-two.

* WebProcess/WebPage/CoordinatedGraphics/CompositingCoordinator.cpp:
(WebKit::CompositingCoordinator::getCoordinatedBuffer):
* WebProcess/WebPage/CoordinatedGraphics/UpdateAtlas.cpp:
(WebKit::UpdateAtlas::UpdateAtlas):
* WebProcess/WebPage/CoordinatedGraphics/UpdateAtlas.h:


Identifier: 195999@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225131 268f45cc-cd09-0410-ab3c-d52691b4dbfc