[GStreamer] Fix USE_GSTREAMER_GL check for GStreamer 1.10 https://bugs.webkit.org/show_bug.cgi?id=180256
Reviewed by Carlos Alberto Lopez Perez.

It's not safe to change the value of build options after WEBKIT_OPTION_END, since it breaks
the option dependency checker. I don't think that actually hurts anything in this case, but
let's not encourage bad practice. Also, it's not good for features to be secretly disabled.
Nobody reads CMake output unless there is a fatal error. Failing the build at the CMake
stage lets distributors know that a feature will be missing. And of course, it's still
possible to disable USE_GSTREAMER_GL to avoid the increased GStreamer requirement.

* Source/cmake/GStreamerChecks.cmake:


Identifier: 196250@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225401 268f45cc-cd09-0410-ab3c-d52691b4dbfc