AlternativePresentationButtonSubstitution::unapply() may not undo substitution https://bugs.webkit.org/show_bug.cgi?id=180279 <rdar://problem/35800599>
Reviewed by Simon Fraser.

Source/WebCore:

Fixes an issue where removing an alternative presentation button substituted for a non-HTML input
element did not restore the original appearance of the element before the substitution.

To substitute the alternative presentation button for a non-HTML input element we attach a
user-agent shadow root to it. Adding a shadow root, including a user-agent shadow root,
tears down the existing renderers for the element. Currently when we unapply such a substitution
we ultimately just remove the shadow root and do not create new renderers for the subtree
that the shadow root was removed from. We need to create new renderers for this subtree
to restore the original appearance of the element before the substitution.

* editing/cocoa/AlternativePresentationButtonSubstitution.cpp:
(WebCore::AlternativePresentationButtonSubstitution::unapply):

LayoutTests:

Update test to ensure that we undo the alternative presentation button substitution
made to an HTML label element.

* fast/forms/alternative-presentation-button/replace-and-remove-expected.html:
* fast/forms/alternative-presentation-button/replace-and-remove.html:

Identifier: 196276@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225431 268f45cc-cd09-0410-ab3c-d52691b4dbfc