Unreviewed, fix byte order macros and address new -Wunknown-pragmas warnings https://bugs.webkit.org/show_bug.cgi?id=180177 <rdar://problem/35774734>
I'm not sure how this code was developed, as it seems to have been designed for GCC, but it
does not use any of GCC's documented byte order macros, and accordingly does not work. Let's
fix it to guarantee there are no problems when building with GCC. I presume it should help
Clang as well.

This time, hopefully without breaking 32-bit macOS, where __BYTE_ORDER actually is defined.

* src/common/third_party/smhasher/src/PMurHash.cpp:
(angle::PMurHash32_Process):

Identifier: 196280@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225436 268f45cc-cd09-0410-ab3c-d52691b4dbfc